+++ src/Data/Eliminator.hs
@@ -25,13 +25,27 @@ Dependently typed elimination functions using @singletons@.
 module Data.Eliminator (
     -- * Eliminator functions
     -- $eliminators
-    elimBool
+    elimAll
+  , elimAny
+  , elimArg
+  , elimBool
+  , elimConst
+  , elimDown
+  , elimDual
   , elimEither
+  , elimFirst
+  , elimIdentity
+  , elimLast
   , elimList
+  , elimMax
   , elimMaybe
+  , elimMin
   , elimNat
   , elimNonEmpty
+  , elimOption
   , elimOrdering
+  , elimProduct
+  , elimSum
   , elimTuple0
   , elimTuple2
   , elimTuple3
@@ -40,15 +54,27 @@ module Data.Eliminator (
   , elimTuple6
   , elimTuple7
   , elimVoid
+  , elimWrappedMonoid
   ) where
 
 import Control.Monad.Extra
 
 import Data.Eliminator.TH
+import Data.Functor.Const (Const(..))
+import Data.Functor.Identity (Identity(..))
 import Data.List.NonEmpty (NonEmpty(..))
+import Data.Monoid hiding (First, Last)
 import Data.Nat
-import Data.Singletons.Prelude
+import Data.Ord (Down(..))
+import Data.Semigroup
+import Data.Singletons.Prelude hiding
+  (All, Any, Const, Last, Min, Max, Product, Sum)
+import Data.Singletons.Prelude.Const (SConst(..))
+import Data.Singletons.Prelude.Identity (SIdentity(..))
 import Data.Singletons.Prelude.List.NonEmpty (SNonEmpty(..))
+import Data.Singletons.Prelude.Monoid hiding (SFirst, SLast)
+import Data.Singletons.Prelude.Ord (SDown(..))
+import Data.Singletons.Prelude.Semigroup
 import Data.Void (Void)
 
 import Language.Haskell.TH.Desugar (tupleNameDegree_maybe)
@@ -69,7 +95,30 @@ The naming conventions are:
   with @~>@ prepended.
 -}
 
-$(concatMapM deriveElim [''Bool, ''Either, ''Maybe, ''Nat, ''NonEmpty, ''Ordering, ''Void])
+$(concatMapM deriveElim
+             [ ''All
+             , ''Any
+             , ''Arg
+             , ''Bool
+             , ''Const
+             , ''Down
+             , ''Dual
+             , ''Either
+             , ''First
+             , ''Identity
+             , ''Last
+             , ''Max
+             , ''Maybe
+             , ''Min
+             , ''Nat
+             , ''NonEmpty
+             , ''Option
+             , ''Ordering
+             , ''Product
+             , ''Sum
+             , ''Void
+             , ''WrappedMonoid
+             ])
 $(deriveElimNamed "elimList" ''[])
 $(concatMapM (\n -> let Just deg = tupleNameDegree_maybe n
                     in deriveElimNamed ("elimTuple" ++ show deg) n)